Allow both service account names and namespaces to be splat
Cherrypicked from
https://github.com/hashicorp/vault-plugin-auth-kubernetes/pull/78 for
allowing wildcards in both service account and namespace for roles.

This limitation is only at API level, but Vault can work perfectly by
just removing the check when populating roles.

Although it is in a plugin, it is a builtin plugin, so we need to
rebuild the vault binary.

This was included in Vault 1.4.X, but since we have Vault 1.2.3 in some
instances, we fork here for the time being.